# perf: add a fast path for converting bytes into array when the bytes has the correct alignment